pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "app-user"
version = "0.1.0"
description = "Auth and user data management for multi-user Python web apps"
requires-python = ">=3.11"
dependencies = [
    "pyjwt>=2.0",
    "pydantic>=2.0",
    "starlette>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]
